js/xpconnect/tests/mochitest/test_bug865260.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=865260
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 865260</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 865260 **/
  SimpleTest.waitForExplicitFinish();
  function go() {
    var exn = "nothrow";
    try { $('ifr').contentWindow['Date']; } catch (e) { exn = e; };
    ok(!!/denied/.exec(exn), "Threw instead of crashing");
    SimpleTest.finish();
  }

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=865260">Mozilla Bug 865260</a>
<p id="display"></p>
<div id="content">
<iframe id="ifr" onload="go();" src="http://example.org/tests/js/xpconnect/tests/mochitest/file_empty.html"></iframe>
</div>
<pre id="test">
</pre>
</body>
</html>
